Migrated anonymous class from SubActionBlueprint into its own AddSubactionEffectsAction class
