+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/actions/blueprints/SubActionBlueprint.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.gempukku.stccg.actions.Action;
 import com.gempukku.stccg.actions.CardPerformedAction;
+import com.gempukku.stccg.actions.turn.AddSubactionEffectsAction;
 import com.gempukku.stccg.actions.turn.SystemQueueAction;
 import com.gempukku.stccg.cards.ActionContext;
 import com.gempukku.stccg.cards.InvalidCardDefinitionException;
@@ -35,33 +36,8 @@ import java.util.List;
 })
 public interface SubActionBlueprint {
     default void addEffectToAction(boolean cost, CardPerformedAction action, ActionContext actionContext) {
-            final SystemQueueAction sysAction = new SystemQueueAction(actionContext.getGame()) {
-                @Override
-                public Action nextAction(DefaultGame cardGame) throws InvalidGameLogicException, PlayerNotFoundException {
-                    try {
-                        // Need to insert them, but in the reverse order
-                        final List<Action> actions = createActions(action, actionContext);
-                        if (actions != null) {
-                            final Action[] effectsArray = actions.toArray(new Action[0]);
-                            for (int i = effectsArray.length - 1; i >= 0; i--)
-                                if (cost)
-                                    action.insertCost(effectsArray[i]);
-                                else
-                                    action.insertEffect(effectsArray[i]);
-                        }
-                    } catch (InvalidCardDefinitionException exp) {
-                        throw new InvalidGameLogicException(exp.getMessage());
-                    }
-                    Action nextAction = getNextAction();
-                    if (nextAction != null)
-                        return nextAction;
-                    else {
-                        setAsSuccessful();
-                        return null;
-                    }
-                }
-            };
-
+            final SystemQueueAction sysAction =
+                    new AddSubactionEffectsAction(actionContext, cost, action, this);
             if (cost) {
                 action.appendCost(sysAction);
             } else {
